# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Meigire OpenApi Demo Documentation",
        "description": "Meigire demo backend test OpenApi description",
        "contact": {
            "email": "meigire0583@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Meigire Demo API Server"
        }
    ],
    "paths": {
...
        "/reqres/login": {
            "post": {
                "tags": [
                    "No. 2"
                ],
                "summary": "Integrasi api dengan handle selain response success - login (nomor 2)",
                "description": "Returns json object. Error object original yang dikembalikan dari reqres.in ada di property original.",
                "operationId": "80b26cee93b5eeb304be1b1dd4602baa",
                "responses": {
                    "500": {
                        "description": "Terjadi error dari reqres"
                    },
                    "200": {
                        "description": "Tidak terjadi error dari reqres"
                    }
                }
            }
        },
        "/reqres/register": {
            "post": {
                "tags": [
                    "No. 2"
                ],
                "summary": "Integrasi api dengan handle selain response success - register (nomor 2)",
                "description": "Returns json object. Error object original yang dikembalikan dari reqres.in ada di property original.",
                "operationId": "81b12db92b15e153b86f633600269237",
                "responses": {
                    "500": {
                        "description": "Terjadi error dari reqres"
                    },
                    "200": {
                        "description": "Tidak terjadi error dari reqres"
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}